#touch coords are now normalized, except for rj scenes which are scaled to 320x320; samplerate changes to 22050 for rj scenes; added patch resource libs & rj soundinput/output which are added to pd search path; abstracting scene loading code, rj scenes detected by .rj ext

+++ src/modules/Osc.h
@@ -31,7 +31,7 @@
 - (void)sendList:(NSArray *)list;
 
 // rj touch event
-- (void)sendTouch:(NSString *)eventType forId:(int)id atX:(int)x andY:(int)y;
+- (void)sendTouch:(NSString *)eventType forId:(int)id atX:(float)x andY:(float)y;
 
 // rj accel event
 - (void)sendAccel:(float)x y:(float)y z:(float)z;
